requirements.txt
# Harmonia Synthesis - Requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.2
google-generativeai==0.3.1
numpy==1.26.2
Pillow==10.1.0
python-docx==1.1.0
tenacity==8.2.3
aiofiles==23.2.1
